refactor: remove size_t from codebase

+++ engine/src/common/global_utils.cc
@@ -9,7 +9,7 @@
 
 #include <stdexcept>
 
-inline void* Alloc32(const size_t size, const bool lazy) {
+inline void* Alloc32(const uint64_t size, const bool lazy) {
     const int flags = MAP_ANONYMOUS | MAP_PRIVATE | (lazy ? 0 : MAP_POPULATE);
 
     void* ptr = mmap(nullptr, size, PROT_READ | PROT_WRITE, flags, -1, 0);
